+++ src/video/v9990/V9990PixelRenderer.cc
@@ -114,7 +114,7 @@ void V9990PixelRenderer::frameEnd(EmuTime::param time)
 		rasterizer->frameEnd(time);
 		auto time2 = Timer::getTime();
 		auto current = time2 - time1;
-		const double ALPHA = 0.2;
+		const float ALPHA = 0.2f;
 		finishFrameDuration = finishFrameDuration * (1 - ALPHA) +
 		                      current * ALPHA;
